+++ Zenref.Ava/Services/SearchTermMessage.cs
@@ -1,0 +1,13 @@
+using System.Collections.ObjectModel;
+
+namespace Zenref.Ava.ViewModels;
+
+public class SearchTermMessage
+{
+    public ObservableCollection<PublicationType> SearchPubCollection { get; init; }
+
+    public SearchTermMessage(ObservableCollection<PublicationType> searchOption)
+    {
+        SearchPubCollection = searchOption;
+    }
+}
