+++ Zenref.Ava/ViewModels/ExportViewModel.cs
@@ -18,17 +18,6 @@ using Zenref.Ava.Views;
 
 namespace Zenref.Ava.ViewModels
 {
-    public class PublicationType
-    {
-        public string Name { get; set; }
-        public ObservableCollection<SearchPublicationType> searchPublicationTypes { get; set; }
-
-        public PublicationType(string name, ObservableCollection<SearchPublicationType> searchPublicationTypes)
-        {
-            Name = name;
-            this.searchPublicationTypes = searchPublicationTypes;
-        }
-    }
     public partial class ExportViewModel : ObservableRecipient, IRecipient<SearchTermMessage>
     {
         /// <summary>
